[pcap2SFlow] new tool pcap2sflow-replay
Aim to read a pcap trace and send it as SFlow Datagram to Skydive
SFlowProbe Agent for example.
